+++ test/api.test.js
@@ -2,7 +2,8 @@
 import test from 'tape';
 import mapboxgl from 'mapbox-gl-js-mock';
 import GLDraw from '../';
-import { createMap, cloneFeature } from './test_utils';
+import createMap from './utils/create_map';
+import getGeoJSON from './utils/get_geojson';
 import AfterNextRender from './utils/after_next_render';
 
 var feature = cloneFeature('point');
@@ -41,7 +42,7 @@ test('API test', t => {
   })
 
   t.test('deleteAll', t => {
-    Draw.add(feature);
+    Draw.add(getGeoJSON('point'));
     Draw.deleteAll();
     t.equals(Draw.getAll().features.length, 0, 'Draw.deleteAll removes all geometries');
     t.end();
@@ -66,16 +67,16 @@ test('API test', t => {
   });
 
   t.test('add', t => {
-    var id = Draw.add(feature)[0];
+    var id = Draw.add(getGeoJSON('point'))[0];
     t.equals(typeof id, 'string', 'valid string id returned on add');
 
     // add featureCollection
-    var listOfIds = Draw.add(cloneFeature('featureCollection'));
-    t.equals(listOfIds.length, cloneFeature('featureCollection').features.length,
+    var listOfIds = Draw.add(getGeoJSON('featureCollection'));
+    t.equals(listOfIds.length, getGeoJSON('featureCollection').features.length,
       'valid string id returned when adding a featureCollection');
     Draw.deleteAll();
 
-    var multiId = Draw.add(cloneFeature('multiPolygon'))[0];
+    var multiId = Draw.add(getGeoJSON('multiPolygon'))[0];
     t.equals('string', typeof multiId, 'accepts multi features');
 
     t.end();
@@ -90,7 +91,7 @@ test('API test', t => {
   });
 
   t.test('add existing feature with changed properties', t => {
-    var id = Draw.add(feature);
+    var id = Draw.add(getGeoJSON('point'));
     var point = Draw.get(id);
 
     setTimeout(function() {
@@ -105,10 +106,10 @@ test('API test', t => {
   });
 
   t.test('get', t => {
-    var id = Draw.add(feature);
+    var id = Draw.add(getGeoJSON('point'));
     var f = Draw.get(id);
     t.deepEquals(
-      feature.geometry.coordinates,
+      getGeoJSON('point').geometry.coordinates,
       f.geometry.coordinates,
       'the geometry added is the same returned by Draw.get'
     );
@@ -118,9 +119,9 @@ test('API test', t => {
   });
 
   t.test('getAll', t => {
-    Draw.add(feature);
+    Draw.add(getGeoJSON('point'));
     t.deepEquals(
-      feature.geometry,
+      getGeoJSON('point').geometry,
       Draw.getAll().features[0].geometry,
       'the geometry added is the same returned by Draw.getAll'
     );
@@ -129,7 +130,7 @@ test('API test', t => {
   });
 
   t.test('delete', t => {
-    var id = Draw.add(feature)[0];
+    var id = Draw.add(getGeoJSON('point'))[0];
     Draw.delete(id);
     t.equals(Draw.getAll().features.length, 0, 'can remove a feature by its id');
     t.end();
