+++ test/draw_point.test.js
@@ -2,7 +2,8 @@ import test from 'tape';
 import mapboxgl from 'mapbox-gl-js-mock';
 import createSyntheticEvent from 'synthetic-dom-events';
 import GLDraw from '../';
-import { click, accessToken, createMap } from './test_utils';
+import click from './utils/mouse_click';
+import createMap from './utils/create_map';
 import makeMouseEvent from './utils/make_mouse_event';
 import CommonSelectors from '../src/lib/common_selectors';
 import drawPointMode from '../src/modes/draw_point';
@@ -160,8 +161,6 @@ test('draw_point render, inactive', t => {
   });
   t.end();
 });
-
-mapboxgl.accessToken = accessToken;
 
 test('draw_point interaction', t => {
   const container = document.createElement('div');
