+++ test/interaction_events.test.js
@@ -4,10 +4,8 @@ import test from 'tape';
 import spy from 'sinon/lib/sinon/spy'; // avoid babel-register-related error by importing only spy
 import createSyntheticEvent from 'synthetic-dom-events';
 import GLDraw from '../';
-import {
-  createMap,
-  click
-} from './test_utils';
+import click from './utils/mouse_click';
+import createMap from './utils/create_map';
 import createAfterNextRender from './utils/after_next_render';
 import makeMouseEvent from './utils/make_mouse_event';
