+++ test/line_string.test.js
@@ -4,14 +4,11 @@ import Feature from '../src/feature_types/feature';
 import LineString from '../src/feature_types/line_string';
 import GLDraw from '../';
 import mapboxgl from 'mapbox-gl-js-mock';
-import {
-  accessToken,
-  createMap,
-  createFeature,
-  getPublicMemberKeys,
-  createMockCtx,
-  drawGeometry
-} from './test_utils';
+import createFeature from './utils/create_feature';
+import getPublicMemberKeys from './utils/get_public_member_keys';
+import createMockCtx from './utils/create_mock_context';
+import drawGeometry from './utils/draw_geometry';
+import createMap from './utils/create_map';
 
 test('LineString constructor and API', t => {
   const rawLine = createFeature('line');
@@ -113,7 +110,6 @@ test('LineString#updateCoordinate', t => {
 
 test('LineString integration', function lineDrawClass(t){
   const lineStringCoordinates = [[0, 0], [1, 1], [2, 2]];
-  mapboxgl.accessToken = accessToken;
   const map = createMap();
   const Draw = GLDraw();
   map.addControl(Draw);
