+++ test/multi_feature.test.js
@@ -5,10 +5,8 @@ import Point from '../src/feature_types/point';
 import Polygon from '../src/feature_types/polygon';
 import LineString from '../src/feature_types/line_string';
 import MultiFeature from '../src/feature_types/multi_feature';
-import {
-  getPublicMemberKeys,
-  createMockCtx
-} from './test_utils';
+import createMockCtx from './utils/create_mock_context';
+import getPublicMemberKeys from './utils/get_public_member_keys';
 
 test('MultiPoint via MultiFeature', t => {
   t.ok(MultiFeature.prototype instanceof Feature, 'inherits from Feature');
