+++ test/utils/create_mock_context.js
@@ -1,0 +1,15 @@
+import spy from 'sinon/lib/sinon/spy'; // avoid babel-register-related error by importing only spy
+
+/**
+ * Returns an mock ctx object with just those properties a Feature
+ * requires.
+ *
+ * @return {Object}
+ */
+export default function createMockCtx() {
+  return {
+    store: {
+      featureChanged: spy()
+    }
+  };
+}
