+++ test/utils/mouse_click.js
@@ -1,0 +1,4 @@
+export default function click(map, payload) {
+  map.fire('mousedown', payload);
+  map.fire('mouseup', payload);
+}
